+++ src/actions/workspaces.ts
@@ -1,12 +1,7 @@
 import { ActionCreator } from 'redux';
 
 import { ExternalLibraryName, Library } from '../components/assessment/assessmentShape';
-import {
-  ICodeDelta,
-  Input,
-  IPlaybackData,
-  PlaybackStatus
-} from '../components/sourcecast/sourcecastShape';
+
 import { IWorkspaceState } from '../reducers/states';
 import * as actionTypes from './actionTypes';
 
@@ -292,144 +287,5 @@ export const updateHasUnsavedChanges: ActionCreator<actionTypes.IAction> = (
   payload: {
     workspaceLocation,
     hasUnsavedChanges
-  }
-});
-
-// SOURCECAST
-export const fetchSourcecastIndex = () => ({
-  type: actionTypes.FETCH_SOURCECAST_INDEX
-});
-
-export const updateSourcecastIndex = (index: any) => ({
-  type: actionTypes.UPDATE_SOURCECAST_INDEX,
-  payload: {
-    index
-  }
-});
-
-export const loadPlaybackData = (playbackData: IPlaybackData) => ({
-  type: actionTypes.LOAD_PLAYBACK_DATA,
-  payload: {
-    playbackData
-  }
-});
-
-export const recordAudioUrl: ActionCreator<actionTypes.IAction> = (audioUrl: string) => ({
-  type: actionTypes.RECORD_AUDIO_URL,
-  payload: {
-    audioUrl
-  }
-});
-
-export const setCodeDeltasToApply: ActionCreator<actionTypes.IAction> = (deltas: ICodeDelta[]) => ({
-  type: actionTypes.SET_CODE_DELTAS_TO_APPLY,
-  payload: {
-    deltas
-  }
-});
-
-export const setInputToApply: ActionCreator<actionTypes.IAction> = (
-  workspaceLocation: WorkspaceLocation,
-  inputToApply: Input
-) => ({
-  type: actionTypes.SET_INPUT_TO_APPLY,
-  payload: {
-    workspaceLocation,
-    inputToApply
-  }
-});
-
-export const setSourcecastData: ActionCreator<actionTypes.IAction> = (
-  title: string,
-  description: string,
-  playbackData: IPlaybackData
-) => ({
-  type: actionTypes.SET_SOURCECAST_DATA,
-  payload: {
-    title,
-    description,
-    playbackData
-  }
-});
-
-export const setSourcecastDuration: ActionCreator<actionTypes.IAction> = (duration: number) => ({
-  type: actionTypes.SET_SOURCECAST_PLAYBACK_DURATION,
-  payload: {
-    duration
-  }
-});
-
-export const setSourcecastStatus: ActionCreator<actionTypes.IAction> = (
-  playbackStatus: PlaybackStatus
-) => ({
-  type: actionTypes.SET_SOURCECAST_PLAYBACK_STATUS,
-  payload: {
-    playbackStatus
-  }
-});
-
-// SOURCEREEL
-export const recordEditorInput: ActionCreator<actionTypes.IAction> = (
-  location: WorkspaceLocation,
-  input: Input
-) => ({
-  type: actionTypes.RECORD_EDITOR_INPUT,
-  payload: {
-    location,
-    input
-  }
-});
-
-export const savePlaybackData = (
-  title: string,
-  description: string,
-  audio: Blob,
-  playbackData: string
-) => ({
-  type: actionTypes.SAVE_PLAYBACK_DATA,
-  payload: {
-    title,
-    description,
-    audio,
-    deltas: playbackData
-  }
-});
-
-export const recordEditorInitValue: ActionCreator<actionTypes.IAction> = (editorValue: string) => ({
-  type: actionTypes.RECORD_EDITOR_INIT_VALUE,
-  payload: {
-    editorValue
-  }
-});
-
-export const timerPause: ActionCreator<actionTypes.IAction> = (timeNow: number) => ({
-  type: actionTypes.TIMER_PAUSE,
-  payload: {
-    timeNow: Date.now()
-  }
-});
-
-export const timerReset = () => ({
-  type: actionTypes.TIMER_RESET
-});
-
-export const timerResume: ActionCreator<actionTypes.IAction> = (timeNow: number) => ({
-  type: actionTypes.TIMER_RESUME,
-  payload: {
-    timeNow: Date.now()
-  }
-});
-
-export const timerStart: ActionCreator<actionTypes.IAction> = (timeNow: number) => ({
-  type: actionTypes.TIMER_START,
-  payload: {
-    timeNow: Date.now()
-  }
-});
-
-export const timerStop: ActionCreator<actionTypes.IAction> = (timeNow: number) => ({
-  type: actionTypes.TIMER_STOP,
-  payload: {
-    timeNow: Date.now()
   }
 });
